% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{addCloned}
\alias{addCloned}
\alias{Subpopulation$addCloned}
\alias{.P$addCloned}
\title{SLiM method addCloned}
\usage{
addCloned(parent)
}
\arguments{
\item{parent}{An object of type Individual object. Must be of length 1 (a
singleton). See details for description.}
}
\value{
An object of type null or Individual object. Return will be of length 1
(a singleton)
}
\description{
Documentation for SLiM function \code{addCloned}, which is a method of the SLiM
class \code{Subpopulation}.
}
\details{
Generates a new offspring individual from the given parent by
clonal reproduction, queues it for addition to the target subpopulation, and
returns it. The new offspring will not be visible as a member of the target
subpopulation until the end of the offspring generation life cycle stage.
The subpopulation of parent will be used to locate applicable mutation() and
modifyChild() callbacks governing the generation of the offspring individual.
Note that this method is only for use in nonWF models. See addCrossed() for
further general notes on the addition of new offspring individuals.
}
