% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{addMutations}
\alias{addMutations}
\alias{Genome$addMutations}
\alias{.G$addMutations}
\title{SLiM method addMutations}
\usage{
addMutations(mutations)
}
\arguments{
\item{mutations}{An object of type Mutation object. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{addMutations}, which is a method of the
SLiM class \code{Genome}.
}
\details{
Add the existing mutations in mutations to the genome, if they are
not already present (if they are already present, they will be ignored), and
if the addition is not prevented by the mutation stacking policy (see the
mutationStackPolicy property of MutationType, section 23.9.1). Calling this
will normally affect the fitness values calculated at the end of the current
generation; if you want current fitness values to be affected, you can call
SLiMSim’s method recalculateFitness() – but see the documentation of that method
for caveats.
}
