% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{addSubpop}
\alias{addSubpop}
\alias{SLiMSim$addSubpop}
\alias{.SS$addSubpop}
\title{SLiM method addSubpop}
\usage{
addSubpop(subpopID, size, sexRatio)
}
\arguments{
\item{subpopID}{An object of type integer or string. Must be of length 1 (a
singleton). See details for description.}

\item{size}{An object of type integer. Must be of length 1 (a singleton). See
details for description.}

\item{sexRatio}{An object of type float. Must be of length 1 (a singleton). The
default value is \code{0.5}. See details for description.}
}
\value{
An object of type Subpopulation object. Return will be of length 1 (a
singleton)
}
\description{
Documentation for SLiM function \code{addSubpop}, which is a method of the SLiM
class \code{SLiMSim}.
}
\details{
Add a new subpopulation with id subpopID and size individuals.
The subpopID parameter may be either an integer giving the ID of the new
subpopulation, or a string giving the name of the new subpopulation (such as
"p5" to specify an ID of 5). Only if sex is enabled in the simulation, the
initial sex ratio may optionally be specified as sexRatio (as the male fraction,
M:M+F); if it is not specified, a default of 0.5 is used. The new subpopulation
will be defined as a global variable immediately by this method (see section
23.13), and will also be returned by this method. Subpopulations added by this
method will initially consist of individuals with empty genomes. In order to
model subpopulations that split from an already existing subpopulation, use
addSubpopSplit().
}
