% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_scripting.R
\name{as.character.slim_script}
\alias{as.character.slim_script}
\title{Convert a slim_script object into text.}
\usage{
\method{as.character}{slim_script}(x)
}
\arguments{
\item{slim_script}{}
}
\value{

}
\description{
Convert a slim_script object into text.
}
