% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{cachedFitness}
\alias{cachedFitness}
\alias{Subpopulation$cachedFitness}
\alias{.P$cachedFitness}
\title{SLiM method cachedFitness}
\usage{
cachedFitness(indices)
}
\arguments{
\item{indices}{An object of type null or integer. See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{cachedFitness}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
The fitness values calculated for the individuals at the indices
given are returned. If NULL is passed, fitness values for all individuals in
the subpopulation are returned. The fitness values returned are cached values;
fitness() callbacks are therefore not called as a side effect of this method.
It is always an error to call cachedFitness() from inside a fitness() callback,
since fitness values are in the middle of being set up. In WF models, it is also
an error to call cachedFitness() from a late() event, because fitness values
for the new offspring generation have not yet been calculated and are undefined.
In nonWF models, the population may be a mixture of new and old individuals,
so instead, NAN will be returned as the fitness of any new individuals whose
fitness has not yet been calculated. When new subpopulations are first created
with addSubpop() or addSubpopSplit(), the fitness of all of the newly created
individuals is considered to be 1.0 until fitness values are recalculated.
}
