% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{containsMutations}
\alias{containsMutations}
\alias{Genome$containsMutations}
\alias{.G$containsMutations}
\alias{Individual$containsMutations}
\alias{.I$containsMutations}
\title{SLiM method containsMutations}
\usage{
containsMutations(mutations)

containsMutations(mutations)
}
\arguments{
\item{mutations}{An object of type Mutation object. See details for description.}
}
\value{
An object of type logical.

An object of type logical.
}
\description{
Documentation for SLiM function \code{containsMutations}, which is a method of
the SLiM class \code{Genome}.

Documentation for SLiM function \code{containsMutations}, which is a method of
the SLiM class \code{Individual}.
}
\details{
Returns a logical vector indicating whether each of the mutations
in mutations is present in the genome; each element in the returned vector
indicates whether the corresponding mutation is present (T) or absent (F). This
method is provided for speed; it is much faster than the corresponding Eidos
code.

Returns a logical vector indicating whether each of the mutations in
mutations is present in the individual (in either of its genomes); each element
in the returned vector indicates whether the corresponding mutation is present
(T) or absent (F). This method is provided for speed; it is much faster than the
corresponding Eidos code.
}
