% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{deregisterScriptBlock}
\alias{deregisterScriptBlock}
\alias{SLiMSim$deregisterScriptBlock}
\alias{.SS$deregisterScriptBlock}
\title{SLiM method deregisterScriptBlock}
\usage{
deregisterScriptBlock(scriptBlocks)
}
\arguments{
\item{scriptBlocks}{An object of type integer or SLiMEidosBlock object. See
details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{deregisterScriptBlock}, which is a method
of the SLiM class \code{SLiMSim}.
}
\details{
All SLiMEidosBlock objects specified by scriptBlocks (either with
SLiMEidosBlock objects or with integer identifiers) will be scheduled for
deregistration. The deregistered blocks remain valid, and may even still be
executed in the current stage of the current generation (see section 24.9); the
blocks are not actually deregistered and deallocated until sometime after the
currently executing script block has completed. To immediately prevent a script
block from executing, even when it is scheduled to execute in the current stage
of the current generation, use the active property of the script block (see
sections 23.10.1 and 24.9).
}
