% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{distanceToPoint}
\alias{distanceToPoint}
\alias{InteractionType$distanceToPoint}
\alias{.IT$distanceToPoint}
\title{SLiM method distanceToPoint}
\usage{
distanceToPoint(individuals1, point)
}
\arguments{
\item{individuals1}{An object of type Individual object. See details for
description.}

\item{point}{An object of type float. See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{distanceToPoint}, which is a method of the
SLiM class \code{InteractionType}.
}
\details{
Returns a vector containing distances between individuals in
individuals1 and the point given by the spatial coordinates in point. The
point vector is interpreted as providing coordinates precisely as specified by
the spatiality of the interaction type; if the interaction type’s spatiality
is "xz", for example, then point\link{0} is assumed to be an x value, and point\link{1}
is assumed to be a z value. Be careful; this means that in general it is
not safe to pass an individual’s spatialPosition property for point, for
example (although it is safe if the spatiality of the interaction matches the
dimensionality of the simulation). A coordinate for a periodic spatial dimension
must be within the spatial bounds for that dimension, since coordinates outside
of periodic bounds are meaningless (pointPeriodic() may be used to ensure this);
coordinates for non-periodic spatial dimensions are not restricted. Importantly,
distances are calculated according to the spatiality of the InteractionType
(as declared in initializeInteractionType()) not the dimensionality of the
model as a whole (as declared in initializeSLiMOptions()). The distances are
therefore interaction distances: the distances that are used to calculate
interaction strengths. If the InteractionType is non-spatial, this method may
not be called. The vector point must be exactly as long as the spatiality of the
InteractionType.
}
