% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{drawByStrength}
\alias{drawByStrength}
\alias{InteractionType$drawByStrength}
\alias{.IT$drawByStrength}
\title{SLiM method drawByStrength}
\usage{
drawByStrength(individual, count)
}
\arguments{
\item{individual}{An object of type Individual object. Must be of length 1 (a
singleton). See details for description.}

\item{count}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type Individual object.
}
\description{
Documentation for SLiM function \code{drawByStrength}, which is a method of the
SLiM class \code{InteractionType}.
}
\details{
Returns up to count individuals drawn from the subpopulation of
individual. The probability of drawing particular individuals is proportional
to the strength of interaction they exert upon individual. This method may be
used with either spatial or non-spatial interactions, but will be more efficient
with spatial interactions that set a short maximum interaction distance. Draws
are done with replacement, so the same individual may be drawn more than once;
sometimes using unique() on the result of this call is therefore desirable. If
more than one draw will be needed, it is much more efficient to use a single
call to drawByStrength(), rather than drawing individuals one at a time. Note
that if no individuals exert a non-zero interaction upon individual, the vector
returned will be zero-length; it is important to consider this possibility.
If the needed interaction strengths have already been calculated, those cached
values are simply used. Otherwise, calling this method triggers evaluation
of the needed interactions, including calls to any applicable interaction()
callbacks.
}
