% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{drawSelectionCoefficient}
\alias{drawSelectionCoefficient}
\alias{MutationType$drawSelectionCoefficient}
\alias{.MT$drawSelectionCoefficient}
\title{SLiM method drawSelectionCoefficient}
\usage{
drawSelectionCoefficient(n)
}
\arguments{
\item{n}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type float or any but object or void or void.
}
\description{
Documentation for SLiM function \code{drawSelectionCoefficient}, which is a
method of the SLiM class \code{MutationType}.
}
\details{
Draws and returns a vector of n selection coefficients using the
currently defined distribution of fitness effects (DFE) for the target mutation
type. See section 23.9 above for discussion of the supported distributions
and their uses. If the DFE is type "s", this method will result in synchronous
execution of the DFE’s script.
}
