% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{getValue}
\alias{getValue}
\alias{GenomicElementType$getValue}
\alias{.GET$getValue}
\alias{Individual$getValue}
\alias{.I$getValue}
\alias{InteractionType$getValue}
\alias{.IT$getValue}
\alias{Mutation$getValue}
\alias{.M$getValue}
\alias{MutationType$getValue}
\alias{.MT$getValue}
\alias{SLiMSim$getValue}
\alias{.SS$getValue}
\alias{Subpopulation$getValue}
\alias{.P$getValue}
\alias{Substitution$getValue}
\alias{.S$getValue}
\title{SLiM method getValue}
\usage{
getValue(key)

getValue(key)

getValue(key)

getValue(key)

getValue(key)

getValue(key)

getValue(key)

getValue(key)
}
\arguments{
\item{key}{An object of type string. Must be of length 1 (a singleton). See
details for description.}
}
\value{
An object of type any but object or void or void or void.

An object of type any but object.

An object of type any but object.

An object of type any but object or void or void or void.

An object of type float or any but object or void or void.

An object of type any but object.

An object of type any but object.

An object of type any but object or void.
}
\description{
Documentation for SLiM function \code{getValue}, which is a method of the SLiM
class \code{GenomicElementType}.

Documentation for SLiM function \code{getValue}, which is a method of the SLiM
class \code{Individual}.

Documentation for SLiM function \code{getValue}, which is a method of the SLiM
class \code{InteractionType}.

Documentation for SLiM function \code{getValue}, which is a method of the SLiM
class \code{Mutation}.

Documentation for SLiM function \code{getValue}, which is a method of the SLiM
class \code{MutationType}.

Documentation for SLiM function \code{getValue}, which is a method of the SLiM
class \code{SLiMSim}.

Documentation for SLiM function \code{getValue}, which is a method of the SLiM
class \code{Subpopulation}.

Documentation for SLiM function \code{getValue}, which is a method of the SLiM
class \code{Substitution}.
}
\details{
Returns the value previously set for the dictionary entry identifier
key using setValue(), or NULL if no value has been set. This dictionary-style
functionality is actually provided by the superclass of GenomicElementType,
SLiMEidosDictionary, although that fact is not presently visible in Eidos since
superclasses are not introspectable.

Returns the value previously set for the dictionary entry identifier
key using setValue(), or NULL if no value has been set. This dictionary-
style functionality is actually provided by the superclass of Individual,
SLiMEidosDictionary, although that fact is not presently visible in Eidos since
superclasses are not introspectable.

Returns the value previously set for the dictionary entry identifier
key using setValue(), or NULL if no value has been set. This dictionary-
style functionality is actually provided by the superclass of InteractionType,
SLiMEidosDictionary, although that fact is not presently visible in Eidos since
superclasses are not introspectable.

Returns the value previously set for the dictionary entry identifier
key using setValue(), or NULL if no value has been set. This dictionary-
style functionality is actually provided by the superclass of Mutation,
SLiMEidosDictionary, although that fact is not presently visible in Eidos since
superclasses are not introspectable.

Returns the value previously set for the dictionary entry identifier
key using setValue(), or NULL if no value has been set. This dictionary-
style functionality is actually provided by the superclass of MutationType,
SLiMEidosDictionary, although that fact is not presently visible in Eidos since
superclasses are not introspectable.

Returns the value previously set for the dictionary entry identifier
key using setValue(), or NULL if no value has been set. This dictionary-
style functionality is actually provided by the superclass of SLiMSim,
SLiMEidosDictionary, although that fact is not presently visible in Eidos since
superclasses are not introspectable.

Returns the value previously set for the dictionary entry identifier
key using setValue(), or NULL if no value has been set. This dictionary-
style functionality is actually provided by the superclass of Subpopulation,
SLiMEidosDictionary, although that fact is not presently visible in Eidos since
superclasses are not introspectable.

Returns the value previously set for the dictionary entry identifier
key using setValue(), or NULL if no value has been set. This dictionary-
style functionality is actually provided by the superclass of Substitution,
SLiMEidosDictionary, although that fact is not presently visible in Eidos since
superclasses are not introspectable.
}
