% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{initializeSex}
\alias{initializeSex}
\alias{Initialize$initializeSex}
\alias{.Init$initializeSex}
\title{SLiM method initializeSex}
\usage{
initializeSex(chromosomeType, xDominanceCoeff)
}
\arguments{
\item{chromosomeType}{An object of type string or numeric. Must be of length 1
(a singleton). See details for description.}

\item{xDominanceCoeff}{An object of type string or numeric. Must be of length 1
(a singleton). The default value is \code{1}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{initializeSex}, which is a method of the
SLiM class \code{Initialize}.
}
\details{
Enable and configure sex in the simulation. The argument chromosomeType
gives the type of chromosome to be simulated; this should be "A", "X", or "Y".
If the chromosomeType is "X", the optional xDominanceCoeff parameter can supply
the dominance coefficient used when a mutation is present in an XY male, and is
thus "heterozygous” (but in a different sense than the heterozygosity of an XX
female with one copy of the mutation). Calling this function has the side effect
of enabling sex in the simulation; individuals will be male and female (rather
than hermaphroditic) regardless of the chromosomeType chosen for simulation.
There is no way to disable sex once it has been enabled; if you don’t want to
have sex, don’t call this function.
}
