% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{initializeSLiMModelType}
\alias{initializeSLiMModelType}
\alias{Initialize$initializeSLiMModelType}
\alias{.Init$initializeSLiMModelType}
\title{SLiM method initializeSLiMModelType}
\usage{
initializeSLiMModelType(modelType)
}
\arguments{
\item{modelType}{An object of type string. Must be of length 1 (a singleton).
See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{initializeSLiMModelType}, which is a
method of the SLiM class \code{Initialize}.
}
\details{
Configure the type of SLiM model used for the simulation. At present,
one of two model types may be selected. If modelType is "WF", SLiM will use a
Wright–Fisher (WF) model; this is the model type that has always been supported
by SLiM, and is the model type used if initializeSLiMModelType() is not called.
If modelType is "nonWF", SLiM will use a non-Wright–Fisher (nonWF) model
instead; this is a new model type supported by SLiM 3.0 and above (see section
1.6). If initializeSLiMModelType() is called at all then it must be called
before any other initialization function, so that SLiM knows from the outset
which features are enabled and which are not.
}
