% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{mmJukesCantor}
\alias{mmJukesCantor}
\alias{SLiMBuiltin$mmJukesCantor}
\alias{.SB$mmJukesCantor}
\title{SLiM method mmJukesCantor}
\usage{
mmJukesCantor(alpha)
}
\arguments{
\item{alpha}{An object of type float. Must be of length 1 (a singleton). See
details for description.}
}
\value{
An object of type string or float or float or float or integer or float
or integer.
}
\description{
Documentation for SLiM function \code{mmJukesCantor}, which is a method of the
SLiM class \code{SLiMBuiltin}.
}
\details{
Returns a mutation matrix representing a Jukes–Cantor (1969) model with
mutation rate alpha to each possible alternative nucleotide at a site: This 2×2
matrix is suitable for use with initializeGenomicElementType(). Note that the
actual mutation rate produced by this matrix is 3*alpha. 0 α α α α 0 α α α α 0 α
α α α 0
}
