% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{mmKimura}
\alias{mmKimura}
\alias{SLiMBuiltin$mmKimura}
\alias{.SB$mmKimura}
\title{SLiM method mmKimura}
\usage{
mmKimura(alpha, beta)
}
\arguments{
\item{alpha}{An object of type float or float. Must be of length 1 (a
singleton). See details for description.}

\item{beta}{An object of type float or float. Must be of length 1 (a singleton).
See details for description.}
}
\value{
An object of type string or float or float or float or integer or float
or integer.
}
\description{
Documentation for SLiM function \code{mmKimura}, which is a method of the SLiM
class \code{SLiMBuiltin}.
}
\details{
Returns a mutation matrix representing a Kimura (1980) model with
transition rate alpha and transversion rate beta: This 2×2 matrix is suitable
for use with initializeGenomicElementType(). Note that the actual mutation rate
produced by this model is alpha+2*beta.
}
