% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{mutationCounts}
\alias{mutationCounts}
\alias{SLiMSim$mutationCounts}
\alias{.SS$mutationCounts}
\title{SLiM method mutationCounts}
\usage{
mutationCounts(subpops, mutations)
}
\arguments{
\item{subpops}{An object of type null or Subpopulation object. See details for
description.}

\item{mutations}{An object of type null or Mutation object. The default value is
\code{NULL}. See details for description.}
}
\value{
An object of type integer.
}
\description{
Documentation for SLiM function \code{mutationCounts}, which is a method of the
SLiM class \code{SLiMSim}.
}
\details{
Return an integer vector with the frequency counts of all of the
Mutation objects passed in mutations, within the Subpopulation objects in
subpops. The subpops argument is required, but you may pass NULL to get
population-wide frequency counts. If the optional mutations argument is NULL
(the default), frequency counts will be returned for all of the active Mutation
objects in the simulation – the same Mutation objects, and in the same order,
as would be returned by the mutations property of sim, in other words. See the
-mutationFrequencies() method to obtain float frequencies instead of integer
counts.
}
