% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{mutationFrequencies}
\alias{mutationFrequencies}
\alias{SLiMSim$mutationFrequencies}
\alias{.SS$mutationFrequencies}
\title{SLiM method mutationFrequencies}
\usage{
mutationFrequencies(subpops, mutations)
}
\arguments{
\item{subpops}{An object of type null or Subpopulation object. See details for
description.}

\item{mutations}{An object of type null or Mutation object. The default value is
\code{NULL}. See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{mutationFrequencies}, which is a method of
the SLiM class \code{SLiMSim}.
}
\details{
Return a float vector with the frequencies of all of the Mutation
objects passed in mutations, within the Subpopulation objects in subpops.
The subpops argument is required, but you may pass NULL to get population-
wide frequencies. If the optional mutations argument is NULL (the default),
frequencies will be returned for all of the active Mutation objects in the
simulation – the same Mutation objects, and in the same order, as would
be returned by the mutations property of sim, in other words. See the -
mutationCounts() method to obtain integer counts instead of float frequencies.
}
