% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{mutationsOfType}
\alias{mutationsOfType}
\alias{Genome$mutationsOfType}
\alias{.G$mutationsOfType}
\alias{SLiMSim$mutationsOfType}
\alias{.SS$mutationsOfType}
\title{SLiM method mutationsOfType}
\usage{
mutationsOfType(mutType)

mutationsOfType(mutType)
}
\arguments{
\item{mutType}{An object of type integer or MutationType object. Must be of
length 1 (a singleton). See details for description.}
}
\value{
An object of type Mutation object.

An object of type Mutation object.
}
\description{
Documentation for SLiM function \code{mutationsOfType}, which is a method of the
SLiM class \code{Genome}.

Documentation for SLiM function \code{mutationsOfType}, which is a method of the
SLiM class \code{SLiMSim}.
}
\details{
Returns an object vector of all the mutations that are of the type
specified by mutType, out of all of the mutations in the genome. If you just
need a count of the matching Mutation objects, rather than a vector of the
matches, use -countOfMutationsOfType(); if you need just the positions of
matching Mutation objects, use -positionsOfMutationsOfType(); and if you are
aiming for a sum of the selection coefficients of matching Mutation objects, use
-sumOfMutationsOfType(). This method is provided for speed; it is much faster
than the corresponding Eidos code.

Returns an object vector of all the mutations that are of the type
specified by mutType, out of all of the mutations that are currently active
in the simulation. If you just need a count of the matching Mutation objects,
rather than a vector of the matches, use -countOfMutationsOfType(). This
method is often used to look up an introduced mutation at a later point
in the simulation, since there is no way to keep persistent references to
objects in SLiM. This method is provided for speed; it is much faster than the
corresponding Eidos code.
}
