% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{nearestInteractingNeighbors}
\alias{nearestInteractingNeighbors}
\alias{InteractionType$nearestInteractingNeighbors}
\alias{.IT}
\alias{$nearestInteractingNeighbors}
\title{SLiM method nearestInteractingNeighbors}
\usage{
nearestInteractingNeighbors(individual, count)
}
\arguments{
\item{individual}{An object of type Individual object. Must be of length 1 (a
singleton). See details for description.}

\item{count}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type Individual object.
}
\description{
Documentation for SLiM function \code{nearestInteractingNeighbors}, which is a
method of the SLiM class \code{InteractionType}.
}
\details{
Returns up to count interacting individuals that are spatially
closest to individual, according to the distance metric of the InteractionType.
More specifically, this method returns only individuals which can exert an
interaction upon the focal individual; it does not include individuals that only
feel an interaction from the focal individual. To obtain all of the interacting
individuals within the maximum interaction distance of individual, simply pass
a value for count that is greater than or equal to the size of individual’s
subpopulation. Note that if fewer than count interacting individuals are within
the maximum interaction distance, the vector returned may be shorter than count,
or even zero-length; it is important to check for this possibility even when
requesting a single neighbor. If only the number of interacting individuals
is needed, use interactingNeighborCount() instead. Note that this method uses
interaction eligibility as a criterion; it will not return neighbors that cannot
exert an interaction upon the focal individual (due to sex-segregation, e.g.).
(It will also never return the focal individual as a neighbor of itself.) To
find all neighbors of the focal individual, whether they can interact with it or
not, use nearestNeighbors().
}
