% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{nearestNeighbors}
\alias{nearestNeighbors}
\alias{InteractionType$nearestNeighbors}
\alias{.IT$nearestNeighbors}
\title{SLiM method nearestNeighbors}
\usage{
nearestNeighbors(individual, count)
}
\arguments{
\item{individual}{An object of type Individual object. Must be of length 1 (a
singleton). See details for description.}

\item{count}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type Individual object.
}
\description{
Documentation for SLiM function \code{nearestNeighbors}, which is a method of
the SLiM class \code{InteractionType}.
}
\details{
Returns up to count individuals that are spatially closest to
individual, according to the distance metric of the InteractionType. To obtain
all of the individuals within the maximum interaction distance of individual,
simply pass a value for count that is greater than or equal to the size of
individual’s subpopulation. Note that if fewer than count individuals are within
the maximum interaction distance, the vector returned may be shorter than count,
or even zero-length; it is important to check for this possibility even when
requesting a single neighbor. Note that this method does not use interaction
eligibility as a criterion; it will return neighbors that could not interact
with the focal individual due to sex-segregation. (It will never return the
focal individual as a neighbor of itself, however.) To find only neighbors
that are eligible to exert an interaction upon the focal individual, use
nearestInteractingNeighbors().
}
