% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{nearestNeighborsOfPoint}
\alias{nearestNeighborsOfPoint}
\alias{InteractionType$nearestNeighborsOfPoint}
\alias{.IT$nearestNeighborsOfPoint}
\title{SLiM method nearestNeighborsOfPoint}
\usage{
nearestNeighborsOfPoint(subpop, point, count)
}
\arguments{
\item{subpop}{An object of type Subpopulation object. Must be of length 1 (a
singleton). See details for description.}

\item{point}{An object of type float. See details for description.}

\item{count}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type Individual object.
}
\description{
Documentation for SLiM function \code{nearestNeighborsOfPoint}, which is a
method of the SLiM class \code{InteractionType}.
}
\details{
Returns up to count individuals in subpop that are spatially closest to
point, according to the distance metric of the InteractionType. To obtain all of
the individuals within the maximum interaction distance of point, simply pass a
value for count that is greater than or equal to the size of subpop. Note that
if fewer than count individuals are within the maximum interaction distance, the
vector returned may be shorter than count, or even zero-length; it is important
to check for this possibility even when requesting a single neighbor.
}
