% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_scripting.R
\name{new_slim_script}
\alias{new_slim_script}
\alias{validate_slim_script}
\title{slim_script constructor and validator}
\usage{
new_slim_script(x, ..., nrow, slim_output = "None", slim_input = "None")

validate_slim_script(slim_script)
}
\arguments{
\item{x}{A named list where each element is a list or vector component of a \code{slim_script},
which includes "block_name", "block_id", "start_gen", "end_gen", "callback", and "code"}

\item{...}{Optional attributes to add to the \code{slim_script} object, as name-value pairs.}

\item{nrow}{The number of rows, required}

\item{slim_output}{Optional \code{slim_output} attribute}

\item{slim_input}{Optional \code{slim_input} attribute}

\item{slim_script}{A \code{slim_script} object}
}
\description{
Constructs or Validates a \code{slim_script} object. This is mostly for internal use.
}
