% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{nucleotideFrequencies}
\alias{nucleotideFrequencies}
\alias{SLiMBuiltin$nucleotideFrequencies}
\alias{.SB$nucleotideFrequencies}
\title{SLiM method nucleotideFrequencies}
\usage{
nucleotideFrequencies(sequence)
}
\arguments{
\item{sequence}{An object of type integer or string. See details for
description.}
}
\value{
An object of type string or float or float or float or integer or float
or integer.
}
\description{
Documentation for SLiM function \code{nucleotideFrequencies}, which is a method
of the SLiM class \code{SLiMBuiltin}.
}
\details{
A convenience function that returns a float vector of length four,
providing the frequencies of occurrences of A / C / G / T nucleotides,
respectively, in the supplied nucleotide sequence. The parameter sequence may be
a singleton string (e.g., "TATA"), a string vector of single characters (e.g.,
"T", "A", "T", "A"), or an integer vector (e.g., 3, 0, 3, 0), using SLiM’s
standard code of A=0, C=1, G=2, T=3.
}
