% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{output}
\alias{output}
\alias{Genome$output}
\alias{.G$output}
\title{SLiM method output}
\usage{
output(filePath, append)
}
\arguments{
\item{filePath}{An object of type null or string. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{append}{An object of type logical. Must be of length 1 (a singleton). The
default value is \code{F}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{output}, which is a method of the SLiM
class \code{Genome}.
}
\details{
Output the target genomes in SLiM’s native format (see section 25.3.1
for output format details). This low-level output method may be used to output
any sample of Genome objects (the Eidos function sample() may be useful for
constructing custom samples, as may the SLiM class Individual). For output of
a sample from a single Subpopulation, the outputSample() of Subpopulation may
be more straightforward to use. If the optional parameter filePath is NULL (the
default), output is directed to SLiM’s standard output. Otherwise, the output
is sent to the file specified by filePath, overwriting that file if append if F,
or appending to the end of it if append is T. See outputMS() and outputVCF() for
other output formats. Output is generally done in a late() event, so that the
output reflects the state of the simulation at the end of a generation.
}
