% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{outputFixedMutations}
\alias{outputFixedMutations}
\alias{SLiMSim$outputFixedMutations}
\alias{.SS$outputFixedMutations}
\title{SLiM method outputFixedMutations}
\usage{
outputFixedMutations(filePath, append)
}
\arguments{
\item{filePath}{An object of type null or string. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{append}{An object of type logical. Must be of length 1 (a singleton). The
default value is \code{F}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{outputFixedMutations}, which is a method
of the SLiM class \code{SLiMSim}.
}
\details{
Output all fixed mutations – all Substitution objects, in other words
(see section 1.5.2) – in a SLiM native format (see section 25.1.2 for output
format details). If the optional parameter filePath is NULL (the default),
output will be sent to Eidos’s output stream (see section 4.2.1). Otherwise,
output will be sent to the filesystem path specified by filePath, overwriting
that file if append if F, or appending to the end of it if append is T.
Mutations which have fixed but have not been turned into Substitution objects
– typically because convertToSubstitution has been set to F for their mutation
type (see section 23.9.1) – are not output; they are still considered to be
segregating mutations by SLiM. In SLiM 3.3 and later, the output format includes
the nucleotides associated with any nucleotide-based mutations; see section
25.1.2. Output is generally done in a late() event, so that the output reflects
the state of the simulation at the end of a generation.
}
