% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{outputMSSample}
\alias{outputMSSample}
\alias{Subpopulation$outputMSSample}
\alias{.P$outputMSSample}
\title{SLiM method outputMSSample}
\usage{
outputMSSample(
  sampleSize,
  replace,
  requestedSex,
  filePath,
  append,
  filterMonomorphic
)
}
\arguments{
\item{sampleSize}{An object of type integer. Must be of length 1 (a singleton).
See details for description.}

\item{replace}{An object of type logical. Must be of length 1 (a singleton). The
default value is \code{T}. See details for description.}

\item{requestedSex}{An object of type string. Must be of length 1 (a singleton).
The default value is \code{"*"}. See details for description.}

\item{filePath}{An object of type null or string. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{append}{An object of type logical. Must be of length 1 (a singleton). The
default value is \code{F}. See details for description.}

\item{filterMonomorphic}{An object of type logical. Must be of length 1 (a
singleton). The default value is \code{F}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{outputMSSample}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Output a random sample from the subpopulation in MS format (see
section 25.2.2 for output format details). Positions in the output will span
the interval \link{0,1}. A sample of genomes (not entire individuals, note) of
size sampleSize from the subpopulation will be output. The sample may be done
either with or without replacement, as specified by replace; the default is to
sample with replacement. A particular sex of individuals may be requested for
the sample, for simulations in which sex is enabled, by passing "M" or "F" for
requestedSex; passing "*", the default, indicates that genomes from individuals
should be selected randomly, without respect to sex. If the sampling options
provided by this method are not adequate, see the outputMS() method of Genome
for a more flexible low-level option. If the optional parameter filePath is NULL
(the default), output will be sent to Eidos’s output stream (see section 4.2.1).
Otherwise, output will be sent to the filesystem path specified by filePath,
overwriting that file if append if F, or appending to the end of it if append
is T. If filterMonomorphic is F (the default), all mutations that are present
in the sample will be included in the output. This means that some mutations may
be included that are actually monomorphic within the sample (i.e., that exist
in every sampled genome, and are thus apparently fixed). These may be filtered
out with filterMonomorphic = T if desired; note that this option means that
some mutations that do exist in the sampled genomes might not be included in the
output, simply because they exist in every sampled genome. See outputSample()
and outputVCFSample() for other output formats. Output is generally done in a
late() event, so that the output reflects the state of the simulation at the end
of a generation.
}
