% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{outputMutations}
\alias{outputMutations}
\alias{SLiMSim$outputMutations}
\alias{.SS$outputMutations}
\title{SLiM method outputMutations}
\usage{
outputMutations(mutations, filePath, append)
}
\arguments{
\item{mutations}{An object of type Mutation object. See details for description.}

\item{filePath}{An object of type null or string. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{append}{An object of type logical. Must be of length 1 (a singleton). The
default value is \code{F}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{outputMutations}, which is a method of the
SLiM class \code{SLiMSim}.
}
\details{
Output all of the given mutations (see section 25.1.3 for output
format details). This can be used to output all mutations of a given mutation
type, for example. If the optional parameter filePath is NULL (the default),
output will be sent to Eidos’s output stream (see section 4.2.1). Otherwise,
output will be sent to the filesystem path specified by filePath, overwriting
that file if append if F, or appending to the end of it if append is T. In SLiM
3.3 and later, the output format includes the nucleotides associated with any
nucleotide-based mutations; see section 25.1.3. Output is generally done in a
late() event, so that the output reflects the state of the simulation at the end
of a generation.
}
