% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{outputUsage}
\alias{outputUsage}
\alias{SLiMSim$outputUsage}
\alias{.SS$outputUsage}
\title{SLiM method outputUsage}
\usage{
outputUsage(void)
}
\arguments{
\item{void}{An object of type . See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{outputUsage}, which is a method of the
SLiM class \code{SLiMSim}.
}
\details{
Output the current memory usage of the simulation to Eidos’s output
stream. The specifics of what is printed, and in what format, should not be
relied upon as they may change from version to version of SLiM. This method
is primarily useful for understanding where the memory usage of a simulation
predominantly resides, for debugging or optimization. Note that it does not
capture all memory usage by the process; rather, it summarizes the memory usage
by SLiM and Eidos in directly allocated objects and buffers. To get the total
memory usage of the running process (either current or peak), use the Eidos
function usage().
}
