% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{outputVCF}
\alias{outputVCF}
\alias{Genome$outputVCF}
\alias{.G$outputVCF}
\title{SLiM method outputVCF}
\usage{
outputVCF(
  filePath,
  outputMultiallelics,
  append,
  simplifyNucleotides,
  outputNonnucleotides
)
}
\arguments{
\item{filePath}{An object of type null or string. Must be of length 1 (a
singleton). The default value is \code{NULL}. See details for description.}

\item{outputMultiallelics}{An object of type logical. Must be of length 1 (a
singleton). The default value is \code{T}. See details for description.}

\item{append}{An object of type logical. Must be of length 1 (a singleton). The
default value is \code{F}. See details for description.}

\item{simplifyNucleotides}{An object of type logical. Must be of length 1 (a
singleton). The default value is \code{F}. See details for description.}

\item{outputNonnucleotides}{An object of type logical. Must be of length 1 (a
singleton). The default value is \code{T}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{outputVCF}, which is a method of the SLiM
class \code{Genome}.
}
\details{
Output the target genomes in VCF format (see sections 25.2.3, 25.2.4,
and 25.3.3 for output format details). The target genomes are treated as pairs
comprising individuals for purposes of structuring the VCF output, so an even
number of genomes is required. This low-level output method may be used to
output any sample of Genome objects (the Eidos function sample() may be useful
for constructing custom samples, as may the SLiM class Individual). For output
of a sample from a single Subpopulation, the outputVCFSample() of Subpopulation
may be more straightforward to use. If the optional parameter filePath is
NULL (the default), output is directed to SLiM’s standard output. Otherwise,
the output is sent to the file specified by filePath, overwriting that file
if append if F, or appending to the end of it if append is T. The parameters
outputMultiallelics, simplifyNucleotides, and outputNonnucleotides affect
the format of the output produced; see sections 25.2.3 and 25.2.4 for further
discussion. See outputMS() and output() for other output formats. Output is
generally done in a late() event, so that the output reflects the state of the
simulation at the end of a generation.
}
