% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{pointInBounds}
\alias{pointInBounds}
\alias{Subpopulation$pointInBounds}
\alias{.P$pointInBounds}
\title{SLiM method pointInBounds}
\usage{
pointInBounds(point)
}
\arguments{
\item{point}{An object of type float. See details for description.}
}
\value{
An object of type logical.
}
\description{
Documentation for SLiM function \code{pointInBounds}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Returns T if point is inside the spatial boundaries of the
subpopulation, F otherwise. For example, for a simulation with "xy"
dimensionality, if point contains exactly two values constituting an (x,y)
point, the result will be T if and only if ((point\link{0}>=x0) & (point\link{0}<=x1) &
(point\link{1}>=y0) & (point\link{1}<=y1)) given spatial bounds (x0, y0, x1, y1). This
method is useful for implementing absorbing or reprising boundary conditions.
This may only be called in simulations for which continuous space has been
enabled with initializeSLiMOptions(). The length of point must be an exact
multiple of the dimensionality of the simulation; in other words, point may
contain values comprising more than one point. In this case, a logical vector
will be returned in which each element is T if the corresponding point in point
is inside the spatial boundaries of the subpopulation, F otherwise.
}
