% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{pointReflected}
\alias{pointReflected}
\alias{Subpopulation$pointReflected}
\alias{.P$pointReflected}
\title{SLiM method pointReflected}
\usage{
pointReflected(point)
}
\arguments{
\item{point}{An object of type float. See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{pointReflected}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Returns a revised version of point that has been brought inside
the spatial boundaries of the subpopulation by reflection. In brief, if
a coordinate of point lies beyond a spatial boundary, that coordinate is
reflected across the boundary, so that it lies inside the boundary by the
same magnitude that it previously lay outside the boundary. This is done
iteratively until all coordinates lie inside the subpopulation’s boundaries.
This method is useful for implementing reflecting boundary conditions. This
may only be called in simulations for which continuous space has been enabled
with initializeSLiMOptions(). The length of point must be an exact multiple of
the dimensionality of the simulation; in other words, point may contain values
comprising more than one point. In this case, each point will be processed as
described above and a new vector containing all of the processed points will be
returned.
}
