% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{pointUniform}
\alias{pointUniform}
\alias{Subpopulation$pointUniform}
\alias{.P$pointUniform}
\title{SLiM method pointUniform}
\usage{
pointUniform(n)
}
\arguments{
\item{n}{An object of type integer. Must be of length 1 (a singleton). The
default value is \code{1}. See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{pointUniform}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Returns a new point (or points, for n > 1) generated from uniform draws
for each coordinate, within the spatial boundaries of the subpopulation. The
returned vector will contain n points, each comprised of a number of coordinates
equal to the dimensionality of the simulation, so it will be of total length
n*dimensionality. This may only be called in simulations for which continuous
space has been enabled with initializeSLiMOptions().
}
