% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{positionsOfMutationsOfType}
\alias{positionsOfMutationsOfType}
\alias{Genome$positionsOfMutationsOfType}
\alias{.G$positionsOfMutationsOfType}
\title{SLiM method positionsOfMutationsOfType}
\usage{
positionsOfMutationsOfType(mutType)
}
\arguments{
\item{mutType}{An object of type integer or MutationType object. Must be of
length 1 (a singleton). See details for description.}
}
\value{
An object of type integer.
}
\description{
Documentation for SLiM function \code{positionsOfMutationsOfType}, which is a
method of the SLiM class \code{Genome}.
}
\details{
Returns the positions of mutations that are of the type specified by
mutType, out of all of the mutations in the genome. If you need a vector of the
matching Mutation objects, rather than just positions, use -mutationsOfType().
This method is provided for speed; it is much faster than the corresponding
Eidos code.
}
