% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{reconstruct.slimr_script}
\alias{reconstruct.slimr_script}
\title{Reconstruct slimrlang code to make this slimr_script}
\usage{
\method{reconstruct}{slimr_script}(x, ...)
}
\arguments{
\item{x}{slimr_script object to reconstruct}

\item{...}{Further arguments, passed to or from other methods.}
}
\value{
A character vector of length one containing the reconstructed code.
}
\description{

}
