% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{relatedness}
\alias{relatedness}
\alias{Individual$relatedness}
\alias{.I$relatedness}
\title{SLiM method relatedness}
\usage{
relatedness(individuals)
}
\arguments{
\item{individuals}{An object of type Individual object. See details for
description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{relatedness}, which is a method of the
SLiM class \code{Individual}.
}
\details{
Returns a vector containing the degrees of relatedness between the
receiver and each of the individuals in individuals. The relatedness between A
and B is always 1.0 if A and B are actually the same individual; this facility
works even if SLiM’s optional pedigree tracking is turned off (in which case
all other relatedness values will be 0.0). Otherwise, if pedigree tracking is
turned on with initializeSLiMOptions(keepPedigrees=T), this method will use
the pedigree information described in section 23.6.1 to construct a relatedness
estimate. More specifically, if information about the grandparental generation
is available, then each grandparent shared by A and B contributes 0.125 towards
the total relatedness, for a maximum value of 0.5 with four shared grandparents.
If grandparental information in unavailable, then if parental information is
available it is used, with each parent shared by A and B contributing 0.25,
again for a maximum of 0.5. If even parental information is unavailable, then
the relatedness is assumed to be 0.0. Again, however, if A and B are the same
individual, the relatedness will be 1.0 in all cases. Note that this relatedness
is simply pedigree-based relatedness. This does not necessarily correspond
to genetic relatedness, because of the effects of factors like assortment and
recombination.
}
