% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setAncestralNucleotides}
\alias{setAncestralNucleotides}
\alias{Chromosome$setAncestralNucleotides}
\alias{.c$setAncestralNucleotides}
\title{SLiM method setAncestralNucleotides}
\usage{
setAncestralNucleotides(sequence)
}
\arguments{
\item{sequence}{An object of type integer or string. See details for
description.}
}
\value{
An object of type integer. Return will be of length 1 (a singleton)
}
\description{
Documentation for SLiM function \code{setAncestralNucleotides}, which is a
method of the SLiM class \code{Chromosome}.
}
\details{
This method, which may be called only in nucleotide-based models
(see section 1.8), replaces the ancestral nucleotide sequence for the
model. The sequence parameter is interpreted exactly as it is in the
initializeAncestralSequence() function; see that documentation for details
(section 23.1). The length of the ancestral sequence is returned. It is unusual
to replace the ancestral sequence in a running simulation, since the nucleotide
states of segregating and fixed mutations will depend upon the original
ancestral sequence. It can be useful when loading a new population state with
readFromMS() or readFromVCF(), such as when resetting the simulation state to an
earlier state in a conditional simulation; however, that is more commonly done
using readFromPopulationFile() with a SLiM or .trees file.
}
