% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setCloningRate}
\alias{setCloningRate}
\alias{Subpopulation$setCloningRate}
\alias{.P$setCloningRate}
\title{SLiM method setCloningRate}
\usage{
setCloningRate(rate)
}
\arguments{
\item{rate}{An object of type numeric. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setCloningRate}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Set the cloning rate of this subpopulation. The rate is changed to
rate, which should be between 0.0 and 1.0, inclusive. Clonal reproduction can
be enabled in both non-sexual (i.e. hermaphroditic) and sexual simulations. In
non-sexual simulations, rate must be a singleton value representing the overall
clonal reproduction rate for the subpopulation. In sexual simulations, rate
may be either a singleton (specifying the clonal reproduction rate for both
sexes) or a vector containing two numeric values (the female and male cloning
rates specified separately, at indices 0 and 1 respectively). During mating and
offspring generation, the probability that any given offspring individual will
be generated by cloning – by asexual reproduction without gametes or meiosis –
will be equal to the cloning rate (for its sex, in sexual simulations) set in
the parental (not the offspring!) subpopulation.
}
