% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setDistribution}
\alias{setDistribution}
\alias{MutationType$setDistribution}
\alias{.MT$setDistribution}
\title{SLiM method setDistribution}
\usage{
setDistribution(distributionType, ...)
}
\arguments{
\item{distributionType}{An object of type string. Must be of length 1 (a
singleton). See details for description.}

\item{...}{An object of type NA. NA See details for description.}
}
\value{
An object of type float or any but object or void or void.
}
\description{
Documentation for SLiM function \code{setDistribution}, which is a method of the
SLiM class \code{MutationType}.
}
\details{
Set the distribution of fitness effects for a mutation type. The
distributionType may be "f", in which case the ellipsis ... should supply a
numeric$ fixed selection coefficient; "e", in which case the ellipsis should
supply a numeric$ mean selection coefficient for the exponential distribution;
"g", in which case the ellipsis should supply a numeric$ mean selection
coefficient and a numeric$ alpha shape parameter for a gamma distribution; "n",
in which case the ellipsis should supply a numeric$ mean selection coefficient
and a numeric$ sigma (standard deviation) parameter for a normal distribution;
"w", in which case the ellipsis should supply a numeric$ λ scale parameter and a
numeric$ k shape parameter for a Weibull distribution; or "s", in which case the
ellipsis should supply a string$ Eidos script parameter. See section 23.9 above
for discussions of these distributions and their uses. The DFE for a mutation
type is normally a constant in simulations, so be sure you know what you are
doing.
}
