% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setGeneConversion}
\alias{setGeneConversion}
\alias{Chromosome$setGeneConversion}
\alias{.c$setGeneConversion}
\title{SLiM method setGeneConversion}
\usage{
setGeneConversion(
  nonCrossoverFraction,
  meanLength,
  simpleConversionFraction,
  bias
)
}
\arguments{
\item{nonCrossoverFraction}{An object of type numeric or numeric or numeric or
numeric. Must be of length 1 (a singleton). See details for description.}

\item{meanLength}{An object of type numeric or numeric or numeric or numeric.
Must be of length 1 (a singleton). See details for description.}

\item{simpleConversionFraction}{An object of type numeric or numeric or numeric
or numeric. Must be of length 1 (a singleton). See details for description.}

\item{bias}{An object of type numeric or numeric or numeric or numeric. Must
be of length 1 (a singleton). The default value is \code{0}. See details for
description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setGeneConversion}, which is a method of
the SLiM class \code{Chromosome}.
}
\details{
This method switches the recombination model to the "double-stranded
break (DSB)” model (if it is not already set to that), and configures the
details of the gene conversion tracts that will therefore be modeled (see
section 1.5.6 for discussion of the "DSB” recombination model). The meanings
and effects of the parameters exactly mirror the initializeGeneConversion()
function; see section 23.1 for details.
}
