% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setGenomicElementType}
\alias{setGenomicElementType}
\alias{GenomicElement$setGenomicElementType}
\alias{.GE$setGenomicElementType}
\title{SLiM method setGenomicElementType}
\usage{
setGenomicElementType(genomicElementType)
}
\arguments{
\item{genomicElementType}{An object of type integer or GenomicElementType
object. Must be of length 1 (a singleton). See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setGenomicElementType}, which is a method
of the SLiM class \code{GenomicElement}.
}
\details{
Set the genomic element type used for a genomic element (see section
1.5.4). The genomicElementType parameter should supply the new genomic element
type for the element, either as a GenomicElementType object or as an integer
identifier. The genomic element type for a genomic element is normally a
constant in simulations, so be sure you know what you are doing.
}
