% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setMigrationRates}
\alias{setMigrationRates}
\alias{Subpopulation$setMigrationRates}
\alias{.P$setMigrationRates}
\title{SLiM method setMigrationRates}
\usage{
setMigrationRates(sourceSubpops, rates)
}
\arguments{
\item{sourceSubpops}{An object of type integer or Subpopulation object. See
details for description.}

\item{rates}{An object of type numeric. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setMigrationRates}, which is a method of
the SLiM class \code{Subpopulation}.
}
\details{
Set the migration rates to this subpopulation from the subpopulations
in sourceSubpops to the corresponding rates specified in rates; in other words,
rates gives the expected fractions of the children in this subpopulation that
will subsequently be generated from parents in the subpopulations sourceSubpops
(see section 21.2.1). This method will only set the migration fractions from
the subpopulations given; migration rates from other subpopulations will be
left unchanged (explicitly set a zero rate to turn off migration from a given
subpopulation). The type of sourceSubpops may be either integer, specifying
subpopulations by identifier, or object, specifying subpopulations directly.
}
