% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setMutationFractions}
\alias{setMutationFractions}
\alias{GenomicElementType$setMutationFractions}
\alias{.GET$setMutationFractions}
\title{SLiM method setMutationFractions}
\usage{
setMutationFractions(mutationTypes, proportions)
}
\arguments{
\item{mutationTypes}{An object of type integer or MutationType object. See
details for description.}

\item{proportions}{An object of type numeric. See details for description.}
}
\value{
An object of type any but object or void or void or void.
}
\description{
Documentation for SLiM function \code{setMutationFractions}, which is a method
of the SLiM class \code{GenomicElementType}.
}
\details{
Set the mutation type fractions contributing to a genomic element type.
The mutationTypes vector should supply the mutation types used by the genomic
element (either as MutationType objects or as integer identifiers), and the
proportions vector should be of equal length, specifying the relative proportion
of mutations that will be drawn from each corresponding type (see section
1.5.4). This is normally a constant in simulations, so be sure you know what you
are doing.
}
