% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setMutationMatrix}
\alias{setMutationMatrix}
\alias{GenomicElementType$setMutationMatrix}
\alias{.GET$setMutationMatrix}
\title{SLiM method setMutationMatrix}
\usage{
setMutationMatrix(mutationMatrix)
}
\arguments{
\item{mutationMatrix}{An object of type float. See details for description.}
}
\value{
An object of type any but object or void or void or void.
}
\description{
Documentation for SLiM function \code{setMutationMatrix}, which is a method of
the SLiM class \code{GenomicElementType}.
}
\details{
Sets a new nucleotide mutation matrix for the genomic
element type. This replaces the mutation matrix originally set by
initializeGenomicElementType(). This method may only be called in nucleotide-
based models.
}
