% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setMutationRate}
\alias{setMutationRate}
\alias{Chromosome$setMutationRate}
\alias{.c$setMutationRate}
\title{SLiM method setMutationRate}
\usage{
setMutationRate(rates, ends, sex)
}
\arguments{
\item{rates}{An object of type numeric. See details for description.}

\item{ends}{An object of type null or integer. The default value is \code{NULL}.
See details for description.}

\item{sex}{An object of type string. Must be of length 1 (a singleton). The
default value is \code{"*"}. See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setMutationRate}, which is a method of the
SLiM class \code{Chromosome}.
}
\details{
Set the mutation rate per base position per generation along the
chromosome. There are two ways to call this method. If the optional ends
parameter is NULL (the default), then rates must be a singleton value that
specifies a single mutation rate to be used along the entire chromosome. If, on
the other hand, ends is supplied, then rates and ends must be the same length,
and the values in ends must be specified in ascending order. In that case, rates
and ends taken together specify the mutation rates to be used along successive
contiguous stretches of the chromosome, from beginning to end; the last position
specified in ends should extend to the end of the chromosome (as previously
determined, during simulation initialization). See the initializeMutationRate()
function for further discussion of precisely how these rates and positions
are interpreted. If the optional sex parameter is "\emph{" (the default), then the
supplied mutation rate map will be used for both sexes (which is the only option
for hermaphroditic simulations). In sexual simulations sex may be "M" or "F"
instead, in which case the supplied mutation rate map is used only for that sex.
Note that whether sex-specific mutation rate maps will be used is set by the
way that the simulation is initially configured with initializeMutationRate(),
and cannot be changed with this method; so if the simulation was set up to use
sex-specific mutation rate maps then sex must be "M" or "F" here, whereas if
it was set up not to, then sex must be "}" or unsupplied here. If a simulation
needs sexspecific mutation rate maps only some of the time, the male and female
maps can simply be set to be identical the rest of the time. The mutation rate
intervals are normally a constant in simulations, so be sure you know what you
are doing. In nucleotide-based models, setMutationRate() may not be called. If
variation in the mutation rate along the chromosome is desired, setHotspotMap()
should be used.
}
