% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setSelectionCoeff}
\alias{setSelectionCoeff}
\alias{Mutation$setSelectionCoeff}
\alias{.M$setSelectionCoeff}
\title{SLiM method setSelectionCoeff}
\usage{
setSelectionCoeff(selectionCoeff)
}
\arguments{
\item{selectionCoeff}{An object of type float. Must be of length 1 (a
singleton). See details for description.}
}
\value{
An object of type any but object or void or void or void.
}
\description{
Documentation for SLiM function \code{setSelectionCoeff}, which is a method of
the SLiM class \code{Mutation}.
}
\details{
Set the selection coefficient of the mutation to selectionCoeff. The
selection coefficient will be changed for all individuals that possess the
mutation, since they all share a single Mutation object (note that the dominance
coefficient will remain unchanged, as it is determined by the mutation type).
This is normally a constant in simulations, so be sure you know what you are
doing; often setting up a fitness() callback (see section 24.2) is preferable,
in order to modify the selection coefficient in a more limited and controlled
fashion (see section 10.5 for further discussion of this point). Changing this
will normally affect the fitness values calculated at the end of the current
generation; if you want current fitness values to be affected, you can call
SLiMSim’s method recalculateFitness() – but see the documentation of that method
for caveats.
}
