% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setSelfingRate}
\alias{setSelfingRate}
\alias{Subpopulation$setSelfingRate}
\alias{.P$setSelfingRate}
\title{SLiM method setSelfingRate}
\usage{
setSelfingRate(rate)
}
\arguments{
\item{rate}{An object of type numeric. Must be of length 1 (a singleton). See
details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setSelfingRate}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Set the selfing rate of this subpopulation. The rate is changed to
rate, which should be between 0.0 and 1.0, inclusive. Selfing can only be
enabled in non-sexual (i.e. hermaphroditic) simulations. During mating and
offspring generation, the probability that any given offspring individual will
be generated by selfing – by self-fertilization via gametes produced by meiosis
by a single parent – will be equal to the selfing rate set in the parental (not
the offspring!) subpopulation (see section 21.2.1).
}
