% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setSexRatio}
\alias{setSexRatio}
\alias{Subpopulation$setSexRatio}
\alias{.P$setSexRatio}
\title{SLiM method setSexRatio}
\usage{
setSexRatio(sexRatio)
}
\arguments{
\item{sexRatio}{An object of type float. Must be of length 1 (a singleton). See
details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setSexRatio}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Set the sex ratio of this subpopulation to sexRatio. As defined
in SLiM, this is actually the fraction of the subpopulation that is male;
in other words, the M:(M+F) ratio. This will take effect when children are
next generated; it does not change the current subpopulation state. Unlike
the selfing rate, the cloning rate, and migration rates, the sex ratio is
deterministic: SLiM will generate offspring that exactly satisfy the requested
sex ratio (within integer roundoff limits). See section 21.2.1 for further
details.
}
