% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{setSubpopulationSize}
\alias{setSubpopulationSize}
\alias{Subpopulation$setSubpopulationSize}
\alias{.P$setSubpopulationSize}
\title{SLiM method setSubpopulationSize}
\usage{
setSubpopulationSize(size)
}
\arguments{
\item{size}{An object of type integer. Must be of length 1 (a singleton). See
details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{setSubpopulationSize}, which is a method
of the SLiM class \code{Subpopulation}.
}
\details{
Set the size of this subpopulation to size individuals. This
will take effect when children are next generated; it does not change the
current subpopulation state. Setting a subpopulation to a size of 0 does
have some immediate effects that serve to disconnect it from the simulation:
the subpopulation is removed from the list of active subpopulations, the
subpopulation is removed as a source of migration for all other subpopulations,
and the symbol representing the subpopulation is undefined. In this case, the
subpopulation itself remains unchanged until children are next generated (at
which point it is deallocated), but it is no longer part of the simulation and
should not be used.
}
