% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{simulationFinished}
\alias{simulationFinished}
\alias{SLiMSim$simulationFinished}
\alias{.SS$simulationFinished}
\title{SLiM method simulationFinished}
\usage{
simulationFinished(void)
}
\arguments{
\item{void}{An object of type . See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{simulationFinished}, which is a method of
the SLiM class \code{SLiMSim}.
}
\details{
Declare the current simulation finished. Normally SLiM ends a
simulation when, at the end of a generation, there are no script events or
callbacks registered for any future generation (excluding scripts with no
declared end generation). If you wish to end a simulation before this condition
is met, a call to simulationFinished() will cause the current simulation to
end at the end of the current generation. For example, a simulation might self-
terminate if a test for a dynamic equilibrium condition is satisfied. Note that
the current generation will finish executing; if you want the simulation to
stop immediately, you can use the Eidos method stop(), which raises an error
condition.
}
