% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_scripting.R
\name{slim_script}
\alias{slim_script}
\title{Create a SLiMR script}
\usage{
slim_script(...)
}
\arguments{
\item{...}{A list of \code{\link{slim_block}} objects comprising a SLiM script (written in slimr code)}
}
\value{
A \code{slim_script} object that can be used with \code{slimr}'s \code{\link{slim_run_script}}
or converted into a text file for use with SLiM directly using \code{\link{as.character}}.
}
\description{
Create a SLiMR script
}
