% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{spatialMapColor}
\alias{spatialMapColor}
\alias{Subpopulation$spatialMapColor}
\alias{.P$spatialMapColor}
\title{SLiM method spatialMapColor}
\usage{
spatialMapColor(name, value)
}
\arguments{
\item{name}{An object of type string or numeric. Must be of length 1 (a
singleton). See details for description.}

\item{value}{An object of type string or numeric. See details for description.}
}
\value{
An object of type string.
}
\description{
Documentation for SLiM function \code{spatialMapColor}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Looks up the spatial map indicated by name, and uses its color-
translation machinery (as defined by the valueRange and colors parameters to
defineSpatialMap()) to translate each element of value into a corresponding
color string. If the spatial map does not have color-translation capabilities,
an error will result. See the documentation for defineSpatialMap() for
information regarding the details of color translation. See the Eidos manual for
further information on color strings.
}
