% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{spatialMapValue}
\alias{spatialMapValue}
\alias{Subpopulation$spatialMapValue}
\alias{.P$spatialMapValue}
\title{SLiM method spatialMapValue}
\usage{
spatialMapValue(name, point)
}
\arguments{
\item{name}{An object of type string or float. Must be of length 1 (a
singleton). See details for description.}

\item{point}{An object of type string or float. See details for description.}
}
\value{
An object of type float.
}
\description{
Documentation for SLiM function \code{spatialMapValue}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Looks up the spatial map indicated by name, and uses its mapping
machinery (as defined by the gridSize, values, and interpolate parameters to
defineSpatialMap()) to translate the coordinates of point into a corresponding
map value. The length of point must be equal to the spatiality of the spatial
map; in other words, for a spatial map with spatiality "xz", point must be
of length 2, specifying the x and z coordinates of the point to be evaluated.
Interpolation will automatically be used if it was enabled for the spatial map.
Point coordinates are clamped into the range defined by the spatial boundaries,
even if the spatial boundaries are periodic; use pointPeriodic() to wrap the
point coordinates first if desired. See the documentation for defineSpatialMap()
for information regarding the details of value mapping. Beginning in SLiM 3.3,
point may contain more than one point to be looked up. In this case, the length
of point must be an exact multiple of the spatiality of the spatial map; for a
spatial map with spatiality "xz", for example, the length of point must be an
exact multiple of 2, and successive pairs of elements from point (elements 0 and
1, then elements 2 and 3, etc.) will be taken as the x and z coordinates of the
points to be evaluated. This allows spatialMapValue() to be used in a vectorized
fashion.
}
