% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{sumOfMutationsOfType}
\alias{sumOfMutationsOfType}
\alias{Genome$sumOfMutationsOfType}
\alias{.G$sumOfMutationsOfType}
\alias{Individual$sumOfMutationsOfType}
\alias{.I$sumOfMutationsOfType}
\title{SLiM method sumOfMutationsOfType}
\usage{
sumOfMutationsOfType(mutType)

sumOfMutationsOfType(mutType)
}
\arguments{
\item{mutType}{An object of type integer or MutationType object. Must be of
length 1 (a singleton). See details for description.}
}
\value{
An object of type float. Return will be of length 1 (a singleton)

An object of type float. Return will be of length 1 (a singleton)
}
\description{
Documentation for SLiM function \code{sumOfMutationsOfType}, which is a method
of the SLiM class \code{Genome}.

Documentation for SLiM function \code{sumOfMutationsOfType}, which is a method
of the SLiM class \code{Individual}.
}
\details{
Returns the sum of the selection coefficients of all mutations that
are of the type specified by mutType, out of all of the mutations in the genome.
This is often useful in models that use a particular mutation type to represent
QTLs with additive effects; in that context, sumOfMutationsOfType() will provide
the sum of the additive effects of the QTLs for the given mutation type. This
method is provided for speed; it is much faster than the corresponding Eidos
code. Note that this method also exists on Individual, for cases in which the
sum across both genomes of an individual is desired.

Returns the sum of the selection coefficients of all mutations that
are of the type specified by mutType, out of all of the mutations in the
genomes of the individual. This is often useful in models that use a particular
mutation type to represent QTLs with additive effects; in that context,
sumOfMutationsOfType() will provide the sum of the additive effects of the
QTLs for the given mutation type. This method is provided for speed; it is much
faster than the corresponding Eidos code. Note that this method also exists on
Genome, for cases in which the sum for just one genome is desired.
}
