% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{takeMigrants}
\alias{takeMigrants}
\alias{Subpopulation$takeMigrants}
\alias{.P$takeMigrants}
\title{SLiM method takeMigrants}
\usage{
takeMigrants(migrants)
}
\arguments{
\item{migrants}{An object of type Individual object. See details for
description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{takeMigrants}, which is a method of the
SLiM class \code{Subpopulation}.
}
\details{
Immediately moves the individuals in migrants to the target
subpopulation (removing them from their previous subpopulation). Individuals in
migrants that are already in the target subpopulation are unaffected. Note that
the indices and order of individuals and genomes in both the target and source
subpopulations will change unpredictably as a side effect of this method. Note
that this method is only for use in nonWF models, in which migration is managed
manually by the model script. In WF models, migration is managed automatically
by the SLiM core based upon the migration rates set for each subpopulation with
setMigrationRates().
}
