% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{treeSeqSimplify}
\alias{treeSeqSimplify}
\alias{SLiMSim$treeSeqSimplify}
\alias{.SS$treeSeqSimplify}
\title{SLiM method treeSeqSimplify}
\usage{
treeSeqSimplify(void)
}
\arguments{
\item{void}{An object of type . See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{treeSeqSimplify}, which is a method of the
SLiM class \code{SLiMSim}.
}
\details{
Triggers an immediate simplification of the tree sequence recording
tables. This method may only be called if tree sequence recording has been
turned on with initializeTreeSeq(). A call to this method will free up memory
being used by entries that are no longer in the ancestral path of any individual
within the current sample (currently living individuals, in other words, plus
those explicitly added to the sample with treeSeqRememberIndividuals()), but it
can also take a significant amount of time. Typically calling this method is not
necessary; the automatic simplification performed occasionally by SLiM should be
sufficient for most models.
}
