% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_lang.R
\name{unevaluate}
\alias{unevaluate}
\alias{InteractionType$unevaluate}
\alias{.IT$unevaluate}
\title{SLiM method unevaluate}
\usage{
unevaluate(void)
}
\arguments{
\item{void}{An object of type . See details for description.}
}
\value{
An object of type void.
}
\description{
Documentation for SLiM function \code{unevaluate}, which is a method of the SLiM
class \code{InteractionType}.
}
\details{
Discards all evaluation of this interaction, for all subpopulations.
The state of the InteractionType is reset to a state prior to evaluation. This
can be useful if the model state has changed in such a way that the evaluation
already conducted is no longer valid. For example, if the maximum distance
or the interaction function of the InteractionType need to be changed with
immediate effect, or if the data used by an interaction() callback has changed
in such a way that previously calculated interaction strengths are no longer
correct, unevaluate() allows the interaction to begin again from scratch. In
WF models, all interactions are automatically reset to an unevaluated state at
the moment when the new offspring generation becomes the parental generation
(at step 4 in the generation cycle; see section 21.4). In nonWF models,
all interactions are automatically reset to an unevaluated state twice per
generation: immediately after reproduction() callbacks have completed (after
step 1 in the generation cycle; see section 22.1), and immediately before
viability/survival selection (before step 4 in the generation cycle; see section
22.4). Given this automatic invalidation, most simulations have no reason to
call unevaluate().
}
