% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slimr_script.R
\name{reconstruct.slimr_script}
\alias{reconstruct.slimr_script}
\title{Reconstruct slimrlang code to make this slimr_script}
\usage{
\method{reconstruct}{slimr_script}(x, ...)
}
\arguments{
\item{x}{slimr_script object to reconstruct}

\item{...}{}
}
\value{
A character vector containing the reconstructed code, each element is a line of code.
}
\description{
This reconstructs a \code{slimrlang} input sequence to regenerate the given \script(slimr_script)
object. This is useful if you want to edit the SLiM script to add additional functionality,
for example, where you want to incorporate the results of \code{slimrlang}'s internal edits, e.g.
such as removing \code{\link{\%.\%}} special operators, etc. It is also useful when the
\code{slimr_script} object has been created from converting a text-based SLiM script, such as when
using \code{\link[slimr]{as.slimr_script}} from the \code{slimr} package on a character variable.
}
